requirements.txt
Flask
pymongo
bcrypt
python-dotenv
PyJWT